app/backend/requirements.txt
aniso8601==10.0.1
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
marshmallow==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2025.2
referencing==0.36.2
rpds-py==0.24.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
